+++ src/models/map-service.js
@@ -2,29 +2,52 @@ import mapProjections from '../data/map-projections';
 import mapServices from '../data/map-services';
 
 class MapService {
-  constructor (name) {
+  constructor (name, { services = [], projections = [] } = {}) {
     this.name = name;
+
+    // Combine user specified projections with our defaults
+
+    this.projections = mapProjections;
+
+    if (Array.isArray(projections)) {
+      this.projections = this.projections.concat(projections);
+    }
+
+    // Combine user specified services with our defaults
+
+    this.services = mapServices;
+
+    if (Array.isArray(services)) {
+      this.services = this.services.concat(services);
+    }
+
     this.configure();
   }
 
   configure () {
-    const service = serviceByName(this.name);
-    this.product = service.product;
-    this.projection = service.projection;
-    this.format = service.format;
-    this.time = service.time;
-    this.resolution = service.resolution;
-    this.tileSize = service.tileSize;
-    this.attribution = service.attribution;
-    this.crs = projectionByName(service.crs);
+    const service = serviceByName(this.services, this.name);
+
+    if (!service) {
+      throw new Error(`Map: Can not find service "${this.name}"`);
+    }
+
+    this.product = service && service.product;
+    this.projection = service && service.projection;
+    this.format = service && service.format;
+    this.time = service && service.time;
+    this.resolution = service && service.resolution;
+    this.tileSize = service && service.tileSize;
+    this.attribution = service && service.attribution;
+    this.crs = service && projectionByName(service.crs);
 
     // Disable prettier here to preserve the strings as human readable
 
-    // prettier-ignore
-    this.projectionResolution = `${this.projection.toUpperCase()}_${this.resolution}`;
+    if (typeof this.projection === 'string' && this.resolution) {
+      // prettier-ignore
+      this.projectionResolution = `${this.projection.toUpperCase()}_${this.resolution}`;
+    }
 
-    // prettier-ignore
-    this.tile = `https://gibs-{s}.earthdata.nasa.gov/wmts/${this.projection}/best/${this.product}/default/${this.time}/${this.projectionResolution}/{z}/{y}/{x}.${this.format}`;
+    this.tile = service && configureTileEndpoint(this, service.tileEndpoint);
   }
 
   setService (name) {
@@ -39,8 +62,8 @@ export default MapService;
  * serviceByName
  */
 
-function serviceByName (name) {
-  return mapServices.find(service => service.name === name);
+function serviceByName (services, name) {
+  return services.find(service => service.name === name);
 }
 
 /**
@@ -49,4 +72,18 @@ function serviceByName (name) {
 
 function projectionByName (name) {
   return mapProjections.find(projection => projection.name === name);
+}
+
+/**
+ * configureTileEndpoint
+ */
+
+function configureTileEndpoint (properties = {}, endpoint) {
+  if (typeof endpoint !== 'string') return '';
+  let tile = endpoint;
+  for (let key in properties) {
+    if (!properties.hasOwnProperty(key)) continue;
+    tile = tile.replace(`{${key}}`, properties[key]);
+  }
+  return tile;
 }