+++ tests/components/OrderList.test.js
@@ -6,36 +6,28 @@ import { OrdersList } from '../../ui';
 describe('Orders List', () => {
   const ordersData = [
     {
-      id: 1,
-      name: 'Two Reams Premium Copy Paper',
-      windowOpen: '04/04/2019',
-      windowClose: '04/08/2019',
-      status: 'accepted'
+      orderId: 1,
+      productIdentifier: 'Two Reams Premium Copy Paper',
+      orderDate: '04/04/2019',
+      orderStatus: 'accepted'
     },
     {
-      id: 2,
-      name: 'One Ream Premium Glossy Paper',
-      windowOpen: '04/04/2019',
-      windowClose: '04/08/2019',
-      status: 'accepted'
+      orderId: 2,
+      productIdentifier: 'One Ream Premium Glossy Paper',
+      orderDate: '04/07/2019',
+      orderStatus: 'accepted'
     },
     {
-      id: 3,
-      name: ' 20 Pads Sticky Squares',
-      windowOpen: '04/04/2019',
-      windowClose: '04/08/2019',
-      status: 'accepted'
+      orderId: 3,
+      productIdentifier: ' 20 Pads Sticky Squares',
+      orderDate: '04/12/2019',
+      orderStatus: 'accepted'
     }
   ];
 
-  const defaultColumns = ['Name', 'Window Open', 'Window Close', 'Status'];
+  const defaultColumns = ['Product', 'Order Date', 'Order Status'];
 
-  const customColumns = [
-    'Orders',
-    'Starting Date',
-    'Ending Date',
-    'Current Status'
-  ];
+  const customColumns = ['Order', 'Date Orders', 'Current Status'];
 
   describe('Render', () => {
     const orderList = shallow(<OrdersList orders={ordersData} />);
@@ -49,10 +41,9 @@ describe('Orders List', () => {
     it('should render a table with the right row values', () => {
       const propRows = orderList.find('Table').prop('rows');
       const firstRow = propRows[0];
-      expect(firstRow[0]).toEqual(ordersData[0].name);
-      expect(firstRow[1]).toEqual(ordersData[0].windowOpen);
-      expect(firstRow[2]).toEqual(ordersData[0].windowClose);
-      expect(firstRow[3]).toEqual(ordersData[0].status);
+      expect(firstRow[0]).toEqual(ordersData[0].productIdentifier);
+      expect(firstRow[1]).toEqual(ordersData[0].orderDate);
+      expect(firstRow[2]).toEqual(ordersData[0].orderStatus);
     });
 
     it('should render a button in the last cell', () => {